vt: check for process == nil before calling kill
If Close is called and the command.Process has not started, a panic can
occur due to directly accessing cmd.Process. Check for Process to exist
before calling Kill.

Reported-by: Koni Marti <koni.marti@gmail.com>

+++ vt.go
@@ -389,7 +389,7 @@ func (vt *VT) scrollDown(n int) {
 func (vt *VT) Close() {
 	vt.mu.Lock()
 	defer vt.mu.Unlock()
-	if vt.cmd != nil {
+	if vt.cmd != nil && vt.cmd.Process != nil {
 		vt.cmd.Process.Kill()
 		vt.cmd.Wait()
 	}
